:art: standardize api response structure

+++ api/housing-service/src/main/java/com/simona/housing/web/controller/HousingController.java
@@ -2,8 +2,11 @@ package com.simona.housing.web.controller;
 
 import com.simona.housing.dto.HousingDto;
 import com.simona.housing.dto.RentalDto;
+import com.simona.housing.dto.ApiResponse;
 import com.simona.housing.model.Housing;
 import com.simona.housing.web.repository.HousingRepository;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -17,6 +20,7 @@ import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
+@Tag(name = "housing")
 @RestController
 public class HousingController {
 
@@ -32,15 +36,22 @@ public class HousingController {
         this.webClient = webClient;
     }
 
+    @Operation(summary = "Find all housings")
     @GetMapping("/housings")
     public Page<Housing> findAll(@RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "20") int size,
                                  @RequestParam(defaultValue = "id") String sortBy) {
         return housingRepository.findAll(PageRequest.of(page, size, Sort.by(sortBy)));
     }
 
+    @Operation(summary = "Find housing by id")
     @GetMapping("/housings/{id}")
-    public HousingDto findById(@PathVariable long id) {
-        Housing foundHousing = housingRepository.findById(id);
+    public ResponseEntity<ApiResponse<HousingDto>> findById(@PathVariable Long id) {
+        if (id == null || id <= 0)
+            return ResponseEntity.badRequest().body(new ApiResponse<>(null, "Id is required"));
+
+        Optional<Housing> foundHousing = housingRepository.findById(id);
+        if (foundHousing.isEmpty())
+            return ResponseEntity.notFound().build();
 
         HousingDto housingDto = new HousingDto();
         List<RentalDto> rentals = webClient.get()
@@ -50,11 +61,15 @@ public class HousingController {
                 .collectList()
                 .block();
 
-        return housingDto.toDto(foundHousing, rentals);
+        return ResponseEntity.ok(new ApiResponse<>(housingDto.toDto(foundHousing.get(), rentals), ""));
     }
 
+    @Operation(summary = "Save a new housing")
     @PostMapping("/housings")
-    public ResponseEntity<Housing> save(@RequestBody Housing housing) {
+    public ResponseEntity<ApiResponse<HousingDto>> save(@RequestBody Housing housing) {
+        if (housing == null)
+            return ResponseEntity.badRequest().body(new ApiResponse<>(null, "Body is required"));
+
         Housing result = housingRepository.save(housing);
 
         if (result == null)
@@ -65,28 +80,37 @@ public class HousingController {
                 .path("/{id}")
                 .buildAndExpand(result.getId())
                 .toUri();
-        return ResponseEntity.created(location).body(result);
+
+        return ResponseEntity.created(location).body(new ApiResponse<>(new HousingDto().toDto(result, null), "" ));
     }
 
+    @Operation(summary = "Update an existing housing")
     @PutMapping("/housings/{id}")
-    public ResponseEntity<Housing> update(
+    public ResponseEntity<ApiResponse<HousingDto>> update(
             @PathVariable Long id, @RequestBody Housing newHousing) {
+        if (id == null || id <= 0)
+            return ResponseEntity.badRequest().body(new ApiResponse<>(null, "Id is required"));
+
         boolean existHousing = housingRepository.existsById(id);
         if (existHousing) {
             newHousing.setId(id);
             Housing updatedHousing = housingRepository.save(newHousing);
-            return ResponseEntity.ok(updatedHousing);
+            return ResponseEntity.ok(new ApiResponse<>(new HousingDto().toDto(updatedHousing, null), "" ));
         } else {
             return ResponseEntity.notFound().build();
         }
     }
 
+    @Operation(summary = "Delete a housing by id")
     @DeleteMapping("housings/{id}")
-    public ResponseEntity<Housing> delete(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<HousingDto>> delete(@PathVariable Long id) {
+        if (id == null || id <= 0)
+            return ResponseEntity.badRequest().body(new ApiResponse<>(null, "Id is required"));
+
         Optional<Housing> existingHousing = housingRepository.findById(id);
         if (existingHousing.isPresent()) {
             housingRepository.deleteById(id);
-            return ResponseEntity.ok(existingHousing.get());
+            return ResponseEntity.ok(new ApiResponse<>(new HousingDto().toDto(existingHousing.get(), null), ""));
         } else {
             return ResponseEntity.notFound().build();
         }
